chore: fix build for custom vitest.d.ts file
