+++ tsconfig.json
@@ -15,15 +15,5 @@
 		"target": "es2022",
 		"types": ["vitest/globals"]
 	},
-	"ts-node": {
-		"esm": true,
-		"files": true,
-		"compilerOptions": {
-			"target": "esnext",
-			"module": "nodenext",
-			"moduleResolution": "nodenext",
-			"esModuleInterop": true,
-			"types": ["vitest/globals"]
-		}
-	}
+	"include": ["./src/vitest.d.ts"]
 }
